Refactor text seach To allow better code sharing between EH and search proc
