+++ src/vs/workbench/services/search/test/node/ripgrepTextSearch.test.ts
@@ -4,46 +4,23 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as assert from 'assert';
-import * as platform from 'vs/base/common/platform';
-import { fixDriveC, fixRegexEndingPattern, getAbsoluteGlob } from 'vs/workbench/services/search/node/ripgrepTextSearch';
+import { fixRegexEndingPattern } from 'vs/workbench/services/search/node/ripgrepTextSearchEH';
 
 suite('RipgrepTextSearch - etc', () => {
-	function testGetAbsGlob(params: string[]): void {
-		const [folder, glob, expectedResult] = params;
-		assert.equal(fixDriveC(getAbsoluteGlob(folder, glob)), expectedResult, JSON.stringify(params));
-	}
-
-	test('getAbsoluteGlob_win', () => {
-		if (!platform.isWindows) {
-			return;
+	test('fixRegexEndingPattern', () => {
+		function testFixRegexEndingPattern([input, expectedResult]: string[]): void {
+			assert.equal(fixRegexEndingPattern(input), expectedResult);
 		}
 
 		[
-			['C:/foo/bar', 'glob/**', '/foo\\bar\\glob\\**'],
-			['c:/', 'glob/**', '/glob\\**'],
-			['C:\\foo\\bar', 'glob\\**', '/foo\\bar\\glob\\**'],
-			['c:\\foo\\bar', 'glob\\**', '/foo\\bar\\glob\\**'],
-			['c:\\', 'glob\\**', '/glob\\**'],
-			['\\\\localhost\\c$\\foo\\bar', 'glob/**', '\\\\localhost\\c$\\foo\\bar\\glob\\**'],
-
-			// absolute paths are not resolved further
-			['c:/foo/bar', '/path/something', '/path/something'],
-			['c:/foo/bar', 'c:\\project\\folder', '/project\\folder']
-		].forEach(testGetAbsGlob);
-	});
-
-	test('getAbsoluteGlob_posix', () => {
-		if (platform.isWindows) {
-			return;
-		}
-
-		[
-			['/foo/bar', 'glob/**', '/foo/bar/glob/**'],
-			['/', 'glob/**', '/glob/**'],
-
-			// absolute paths are not resolved further
-			['/', '/project/folder', '/project/folder'],
-		].forEach(testGetAbsGlob);
+			['foo', 'foo'],
+			['', ''],
+			['^foo.*bar\\s+', '^foo.*bar\\s+'],
+			['foo$', 'foo\\r?$'],
+			['$', '\\r?$'],
+			['foo\\$', 'foo\\$'],
+			['foo\\\\$', 'foo\\\\\\r?$'],
+		].forEach(testFixRegexEndingPattern);
 	});
 
 	test('fixRegexEndingPattern', () => {
