+++ src/vs/workbench/services/search/test/node/textSearch.integrationTest.ts
@@ -11,9 +11,10 @@ import { TPromise } from 'vs/base/common/winjs.base';
 import { FileWalker } from 'vs/workbench/services/search/node/fileSearch';
 import { ISerializedFileMatch, IRawSearch, IFolderSearch } from 'vs/workbench/services/search/node/search';
 import { Engine as TextSearchEngine } from 'vs/workbench/services/search/node/textSearch';
-import { RipgrepEngine } from 'vs/workbench/services/search/node/ripgrepTextSearch';
+import { TextSearchEngineAdapter } from 'vs/workbench/services/search/node/textSearchAdapter';
 import { TextSearchWorkerProvider } from 'vs/workbench/services/search/node/textSearchWorkerProvider';
 import { getPathFromAmdModule } from 'vs/base/common/amd';
+import { CancellationTokenSource } from 'vs/base/common/cancellation';
 
 function countAll(matches: ISerializedFileMatch[]): number {
 	return matches.reduce((acc, m) => acc + m.numMatches, 0);
@@ -62,12 +63,12 @@ function doLegacySearchTest(config: IRawSearch, expectedResultCount: number | Fu
 
 function doRipgrepSearchTest(config: IRawSearch, expectedResultCount: number | Function): TPromise<void> {
 	return new TPromise<void>((resolve, reject) => {
-		let engine = new RipgrepEngine(config);
+		let engine = new TextSearchEngineAdapter(config);
 
 		let c = 0;
-		engine.search((result) => {
-			if (result) {
-				c += result.numMatches;
+		engine.search(new CancellationTokenSource().token, (results) => {
+			if (results) {
+				c += results.reduce((acc, cur) => acc + cur.numMatches, 0);
 			}
 		}, () => { }, (error) => {
 			try {
@@ -75,7 +76,7 @@ function doRipgrepSearchTest(config: IRawSearch, expectedResultCount: number | F
 				if (typeof expectedResultCount === 'function') {
 					assert(expectedResultCount(c));
 				} else {
-					assert.equal(c, expectedResultCount, 'rg');
+					assert.equal(c, expectedResultCount, `rg ${c} !== ${expectedResultCount}`);
 				}
 			} catch (e) {
 				reject(e);
